requirements.txt
psycopg2
discord.py
discord.py[voice]
asyncio
pillow
requests